CacheTable: hopefully made daysInMonth a little faster (comments)
Moved the feb case to the top as most of the time it is a quick check month == 2.  If that fails we move to the 30 day month check and otherwise just return 31 days.  This also gets rid of the ugly error case.  It currently isn’t possible to have any month not 1..12 anyway.