CacheTable: fixed negative days completely (comments)
So now the very first entry is invalid as it would never be checked (first is: abs -1).  Then we start in Feb of 2000, not 1999.  And finally, everything is reversed as it should be because “larger” numbers are actually going backward since they’re negative.  And finally, removed the TODO comment as the issue is solved.

But looking at this, makes me wonder if the table is worth it or perhaps we should just have days go forward to infinity and negDays go back to 1600.  I think it might still be worth it due to how haskell works.  This will create a very small table, if it is ever forced, while an infinite table will be in memory for how ever far forward/backward the user goes.