Added decodeDate’ (comments)
This version will decode the DateTime object without the cache table.  It is intended to be used when one does not want to pay the memory cost for the cache table.  Otherwise there will be decodeDate that uses the table and is faster.

NOTE: There are some bugs in this implementation, e.g. leap days