Added newtype Date for date only values and refactored (comments)
We can just use the existing DateTime object and ignore the other two fields.  Alternatively we can make a new newtype object that just wraps an Int.  If all this doesn’t cause any problem then we might switch to this later.

Also refactored the code to split out the days calculation from the time calculation.