Removed Torsor and refactored (comments)
I decided the type class approach wasn’t the best here because the types don’t line up exactly.  For a duration you can add two durations together and subtract two durations from each other.  For an Instant, on the other hand, you can add a Duration to it, subtract two instances to get a Duration or subtract a Duration from an Instance to get an Instance.

So because of this the type class didn’t really bring much.  I just changed the math symbols to be meaningful names (module name dependant though, add in Instant is different than add in Duration).