Internal: big refactoring (comments)
I’ve gotten rid of the Type module and replaced it with a bunch of Internal modules.  This may seem excessive but this allows me to put the types close to the module they are a type for without worrying about circular references (I hope).  It also means that when I start adding Type Class instances for these types they can be sensible defined.  The instances need to be defined in the same file as the type or they will be orphan instances.  But I can’t do this in a *.Type module because the instances will rely on functions specific to the underlying module.  And I can’t put the types in the top level module because there are inter-dependancies between the different modules.