Olson: switched handling to Either (comments)
I didn’t want to make it specific, but let the usage context decide but apparently that doesn’t work anymore.  Now the fail function in Either just calls error, so my strategy of using returning a Monad would only work for List and Maybe but throw an exception if the caller tried to use an Either.

Because of this we’ll just explicitly use Either at the lower levels and decide later how it bubbles up.