OlsonTest: removed as I have proven correctness (comments)
This was always going to be a pain because, obviously, Olson is a private package.  So I had to move things around to be able to test it, but that was also a lot of trouble.  Instead I got into GHCI and verified that the olson package on package got the same answers as my library.  It does so I can go forward for now without specific tests.  I may have to reevaluate having a private test package later to be able to prove these private packages.