ZoneDateTime: converted transitions to Instant keys (comments)
I changed ZonedDateTime Transitions to be Maps now instead of IntMaps and the key is an instant instead of an integer.  This is because the way that Transitions will be used is the ZonedDateTime will look for an instant close to its internal instant.  We don’t have to have to change ZonedDateTime Instants into integers first.

Also corrected direction of addTransition to match normal Map insert convention and fixed variable names