Instant - bugfix in difference function (comments)
The difference function was crashing any time the difference was less than a second.  To fix this I convert secs and nsecs to Int first to prevent truncation if we go negative, do the math on all values and then normalise.  We need to setup some tests around this to make sure it always works.