LocalTime: changed accessor style (comments)
Got rid of the accessors and made full lenses for each time component.  Also had to update Instant tests and added tests for LocalTime to verify the behaviour.

NOTE: The tests are not catching a problem that exists right now: times normalise (e.g. 22:57:00 with a modify (+5) minutes applied to it will be 23:02:00) but the tests aren’t detecting it.