InstantTest: fix for error message (comments)
I saw an error once where minutes was 25 and 26 was expected but since the previous test lost all the other values, I have no way of knowing when that actually was to reproduce.  This way I will at least get the exact time for testing.

Also removed get definition and included the Util one