Some cleanup (comments)
Updated code using calendars to ensure they have the type variable and the correct name.  Also changing strategy of how UTC is handled.  Now the strategy will be an empty zone but otherwise normal instead of how it was going to be handled with a separate constructor that has to be tested everywhere.  Also did some renaming to make the names a bit nicer while preparing how to load time zone data.