Olson: removed code for isStd and isGmt (comments)
After much searching, it appears the tt_isstd and tt_isgmt fields are not relevant.  They seem to record how the data was specified in the source file but according to the man page, the transition times are always in UTC in unix epoch seconds.  So that being the case, these fields have no impact on what we’re doing so they’re dropped.