TimeZone.Internal: added CalDateTransitionMap (comments)
Added finger tree package to our dependancies and make use of it to store the interval of local times for a given time zone.  This should allow the calendar date time to find the correct transition, detect duplicates and missing time.  This is, as of now, completely untested.

Also had to make some fixes related to the updated stack resolver.  It seems my testing was as little loose with importing.  This is not completely cleaned up yet, just enough to get it to compile.