TimeZone: completed implementation (comments)
The top level implementation is complete for TimeZones.  The Mac platform still has some things that need to be fixed (error handling and actually loading leap seconds) but the TimeZone interface itself should be finished.  The only thing remaining will be to re-export types, classes and functions that will be needed by clients of the module.