Instant: cleanup (comments)
Instant had an issue that it has a circular dependancy on Duration.  The way I was dealing with this was letting modules import some things from Instant instead of Instant.Internal.  I have now corrected this by simply moving the definition for Duration into Instant.Internal.  Now the circular dependancy is gone and the interface has been cleaned up as a result.