Olson: added initial UTC “big bang” transition (comments)
After loading the UTC zone from the file, it was obvious that the calDateMap had good behaviour (one large interval covering the default, and only, ttinfo) but utcMap was just empty.  As a result I’ve used the new “big bang” instant (the smallest possible instant) to make sure we always have an initial ttinfo in effect.