TimeZone: changed error handling strategy (comments)
After discovering how to encode the errors as I wanted, I realised that this strategy would be forcing nested do statements all over my code as well as any client code.  So as a result, my new strategy is that pure code uses MonadThrow to signal error but anything in IO just uses IO.  This makes the actual code much simpler as well as any client code that wants to use it.  We still throw exceptions in case someone wants to catch them.