Instant: changed inUtc to inTimeZone (comments)
Due to how the interface has to work it is not practical to offer any time zone convenience function.  The user will have to get a time zone some other way and pass it to this function.  Note that if we ignored leap seconds then UTC could be created via a pure interface which would allow us to use it here.