ZonedDateTime: added fromCalendarDateTimeAll (comments)
Made a bunch of changes to support this new function that gets all possible ZonedDateTimes for a given CalendarDateTime in a given TimeZone.  In doing so I realised that keeping the CalendarDateTime in local time instead of UTC meant I couldn’t tell the difference between ambiguous times.  To fix this, I added an active TransitionInfo to the ZonedDateTime.  The code is updated in all the relevant areas for this change.  This included changing TimeZone.Platform.fixedOffsetZone to also return the fixed TransitionInfo it uses.