Olson: getCh and getTZString cleanup (comments)
In most of the cases where we were doing getInt8 we actually wanted a character.  We wouldn’t have to for the versions, but since the version is actually stored in character form it’s much clearer to show it that way.

Also fixed the getTZString function function to only consume exactly what it needs instead of slurping up the rest of the file and then potentially not using anything past the newline (there should be nothing after it, but we can’t tell if we eat the rest of the file).