ZoneDateTime: fixed fromCalendarDateTimeLeniently (comments)
To fix this function I got rid of nextCalDateTransition and replaced it with aroundCalDateTransition, which returns the transitions on either side of the requested time.  This allows the resolve skip function to actually be passed the right dates and  fromCalendarDateTimeLeniently to use the offsets of each to calculate how far the time must move into the future to be correct.