TimeZone.Olson: Partial handling of POSIX tz string (comments)
Added code to handle POSIX tz string and did some cleanup.  Cleanup included fixing order of imports and remove unnecessary wrapping of some data which had to be unpacked again for use, which left ugly case statements which use error on the case that should actually be impossible.  Since it is impossible, the code now uses the only valid type there and only wraps in the more general type when it has to.

The strategy for handling the POSIX string is to calculate when the first change would happen, which is most likely in the year following the last transition since the last transition will be switching back to standard time and the first (NOTE: currently we have a bug here), and once we have this time we just add that date to the UTC map with the expression transition info.  For the calDate interval map we add a fake entry for the last date to 1 second before this (offset adjusted) time to close that transition info, and add a file one from this time to Largest with our expression transition info.