Olson: bug fix on addLastMapEntries (comments)
Added a new convenience method for converting instants into day, month, year.  It had to go in Gregorian because Gregorian already depends on Instant so we can’t have a recursive dependancy.  Added another method to TimeZone internal to let the client compute the year first, as we do here.  I hate the dependancy from Olson to Gregorian but the fact is that it does depend on Gregorian for that one function.