TimeZone: major refactor (comments)
Testing against my new C++ utility demonstrated a bug with the structure of how timezone transitions were stored.  To fix it, I’ve moved timezone transition expressions out of the various maps and hold them in a separate type.  The code now checks to see if requested transition is covered by the expression or not.  If not, the map is used, otherwise we use the expression.  This has simplified the code a great deal overall.  I also took the opportunity to make ZonedDateTime respect the interface to TimeZone more (stop pattern matching on it) and made the interface it uses not require knowledge of the internal types to TimeZone.

NOTE: This is not complete, there are still undefineds around that will crash.