Olson: changed calDate calculation to use nanosecond (comments)
Previously we were using the next transition time, minus one second to get the time that the the previous transition ended, but our resolution is nanosecond, therefor we should be setting the time to minus one nanosecond.  Otherwise we leave nearly a second long gap between every transition.

Had to move fromNanoseconds into Duration.Internal to do this