Removed leap seconds handling (comments)
For the moment, all leap seconds are doing is complicating the interface.  It’s not clear how they would be handled but if we bring them back it will probably be with some new type designed for leap seconds.

NOTE: This also makes fixed offset zones considerably easier to create.