Offset: large refactoring (comments)
Changed the TransitionInfo type to take an Offset.  This required a lot of other changes to avoid looping dependancies.  Renamed the OffsetDateTime.Internal to just be Offset.Internal as the former didn’t do anything needed by other modules but the latter does.

Also changed the Offset add/minus functions to be addClamped and minusClamped to make it more obvious their behaviour.

For the various timezone functions, we ignore the Offset smart constructors.  This is partially for speed but also because we don’t want to be clamped (though that should never happen)