BIG CHANGE (comments)
Ok, got the parsing behavior to work for localtime with an intermediate type that gets copied into a real local time after the fact.  It also uses the smart constructor to detect improperly created times (currently impossible).  This is very much incomplete, as we are in the middle of a major refactoring.